Crate bananagrams_solver

Structs

AppState ğŸ”’

Controls the state of the app

Board ğŸ”’

A thin wrapper for handling the board

GameState ğŸ”’

The previous game state

PlayableWords ğŸ”’

Struct returned when getting playable words

Solution ğŸ”’

Struct returned when a board is solved

Enums

Direction ğŸ”’

Enumeration of the direction a word is played

LetterComparison ğŸ”’

For comparing a current hand of letters to a previous hand

LetterUsage ğŸ”’

Enumeration of how many letters have been used

Constants

BOARD_SIZE ğŸ”’

Number rows/columns in the board

EMPTY_VALUE ğŸ”’

Value of an empty cell on the board

MAX_WORD_LENGTH ğŸ”’

The maximum length of any word in the dictionary

REGULAR_TILES ğŸ”’

The number of each letter present in regular Bananagrams

UPPERCASE ğŸ”’

All uppercase letters in the Latin alphabet

Functions

_board_to_string ğŸ”’

Converts a board to a String

board_to_vec ğŸ”’

Converts a board to a vector of vectors of chars

check_filter_after_play ğŸ”’

Checks which words can be played after the first

convert_array_to_word ğŸ”’

Converts a numeric vector representation into a String

convert_word_to_array ğŸ”’

Converts a word into a numeric vector representation

get_playable_words ğŸ”’

Async command executed by the frontend to get the playable words for a given hand of letters

get_previous_idxs ğŸ”’

Gets which indices overlap between previous_play_sequence and new_play_sequence

get_random_letters ğŸ”’

Generates random letters based on user input

is_board_valid_horizontal ğŸ”’

Checks that a board is valid after a word is played horizontally, given the specified list of valid_words Note that this does not check if all words are contiguous; this condition must be enforced elsewhere.

is_board_valid_vertical ğŸ”’

Checks that a board is valid after a word is played vertically, given the specified list of valid_words Note that this does not check if all words are contiguous; this condition must be enforced elsewhere.

is_makeable ğŸ”’

Checks whether a word can be made using the given letters

main ğŸ”’

play_bananagrams ğŸ”’

Async command executed by the frontend to solve a Bananagrams board

play_existing ğŸ”’

Attempts to play off an existing board

play_further ğŸ”’

Recursively solves Bananagrams

play_one_letter ğŸ”’

Tries to play a single letter on the board

play_word ğŸ”’

Plays a word on the board

reset ğŸ”’

Async command executed by the frontend to reset the Banangrams board

undo_play ğŸ”’

Undoes a play on the board

Type Aliases

Letters ğŸ”’

Represents a hand of letters

PlaySequence ğŸ”’

Represents the previous sequence of plays

Word ğŸ”’

A numeric representation of a word