Function bananagrams_solver::get_removable_indices

pub(crate) fn get_removable_indices(
    board: &Board,
    min_col: usize,
    max_col: usize,
    min_row: usize,
    max_row: usize,
) -> Vec<(Vec<(usize, usize)>, usize, usize, usize, usize)>

Gets a vector of vectors of each part of a word that can be validly removed from the board

Arguments

board - The board to check for removable word parts

min_col - The minimum played column

max_col - The maximum played column

min_row - The minimum played row

max_row - The maximum played row

Returns

Vec - Vector of length-5 tuples of (vector of length-2 index tuples of the indices of board that can be validly removed, new_min_col, new_max_col, new_min_row, new_max_row)