Function bananagrams_solver::is_board_valid_vertical

pub(crate) fn is_board_valid_vertical(
    board: &Board,
    min_col: usize,
    max_col: usize,
    min_row: usize,
    max_row: usize,
    start_row: usize,
    end_row: usize,
    col: usize,
    valid_words: &HashSet<&Vec<usize>>,
) -> bool

Checks that a board is valid after a word is played vertically, given the specified list of valid_words Note that this does not check if all words are contiguous; this condition must be enforced elsewhere.

Arguments

board - Board being checked

min_col - Minimum x (column) index of the subsection of the board to be checked

max_col - Maximum x (column) index of the subsection of the board to be checked

min_row - Minimum y (row) index of the subsection of the board to be checked

max_row - Maximum y (row) index of the subsection of the board to be checked

start_row - Starting row of the word played

end_row - Ending row of the word played

col - Column of the word played

valid_words - HashSet of all valid words as Vec<usize>s

Returns

bool - whether the given board is made only of valid words