Function bananagrams_solver::try_play_word_vertically

pub(crate) fn try_play_word_vertically(
    board: &mut Board,
    word: &Vec<usize>,
    min_col: usize,
    max_col: usize,
    min_row: usize,
    max_row: usize,
    valid_words_vec: &Vec<&Vec<usize>>,
    valid_words_set: &HashSet<&Vec<usize>>,
    letters: [usize; 26],
    depth: usize,
    words_checked: &mut usize,
    letters_on_board: &mut [usize; 26],
    filter_letters_on_board: usize,
    max_words_to_check: usize,
    stop_t: &Arc<AtomicBool>,
) -> Result<Option<(bool, usize, usize, usize, usize)>, ()>

Tries to play a word vertically anywhere on the board

Arguments

board - The Board to modify in-place

word - Word to try to play

min_col - Minimum occupied column index in board

max_col - Maximum occupied column index in board

min_row - Minimum occupied row index in board

max_row - Maximum occupied row index in board

valid_words_vec - Vector of vectors, each representing a word (see convert_word_to_array)

valid_words_set - HashSet of vectors, each representing a word (a HashSet version of valid_words_vec for faster membership checking)

letters - Length-26 array of the number of each letter in the hand

depth - Depth of the current recursive call

words_checked - The number of words checked in total

letters_on_board - Length-26 array of the number of each letter currently present on the board

filter_letters_on_board - Maximum number of letters currently on the board that can be used in a newly played word

max_words_to_check - Maximum number of words to check before stopping

stop_t - AtomicBool that, when set, indicates that processing should stop

Returns

Result with Option upon success with:

bool - Whether the word could be validly played

usize - Minimum occupied column index in board

usize - Maximum occupied column index in board

usize - Minimum occupied row index in board

usize - Maximum occupied row index in board

or None if no valid playing location was found, or empty Err if another thread signalled to stop